feat: merge icons.ts into theme.ts, deprecate icons.ts wrapper (WL-0MSJ4BT4Z009LIB4)
Move all icon content (maps, functions, IconOptions interface) from
src/icons.ts into src/theme.ts, making it the single source of truth
for colours and icons.

Replace src/icons.ts with a deprecated re-export wrapper that forwards
all 28 public API names to theme.ts.

Update tests to import from theme.ts and add wrapper parity tests.
Update docs/icons-design.md to reflect the new module layout.
