+++ docs/icons-design.md
@@ -7,6 +7,10 @@
 > [dcb09ac](https://github.com/TheWizardsCode/ContextHub/commit/dcb09ac),
 > [f3ca18b](https://github.com/TheWizardsCode/ContextHub/commit/f3ca18b)
 
+> **Note:** As of WL-0MSJ4BT4Z009LIB4, all icon content is now in
+> `src/theme.ts`.  `src/icons.ts` is a deprecated re-export wrapper that will
+> be removed in a future release.
+
 ## Overview
 
 This document defines the icon set for work item **priority** and **status** used
@@ -254,11 +258,13 @@ The icon lookup is a simple `Map<string, string>` or plain object lookup —
 O(1) per call, negligible runtime cost. No SVG, image loading, or network
 requests are involved.
 
-**Design decision:** Create a single `src/icons.ts` module that exports pure
-functions:
+**Design decision:** Create a single module (`src/theme.ts`) that exports pure
+functions. (As of WL-0MSJ4BT4Z009LIB4, icon definitions were merged into
+`src/theme.ts` alongside the chalk colour theme; `src/icons.ts` is now a
+@deprecated re-export wrapper.)
 
 ```ts
-// src/icons.ts
+// src/theme.ts
 
 export interface IconOptions {
   /** When true, use text fallback instead of emoji/icon glyph */
@@ -296,7 +302,13 @@ export function iconsEnabled(opts?: { noIcons?: boolean }): boolean;
 
 ## 13. Implementation Guide
 
-### 13.1 CLI List Rendering
+### 13.1 Module Layout
+
+Icon maps and functions are defined in `src/theme.ts` alongside the chalk colour
+theme.  The former `src/icons.ts` is a deprecated re-export wrapper that
+re-exports everything from `src/theme.js`.
+
+### 13.2 CLI List Rendering
 
 Icons are prepended before the title in CLI list output:
 
@@ -311,7 +323,7 @@ Icons are prepended before the title in CLI list output:
 > The `dist/icons.ts` functions remain in use by the CLI and the retained
 > session-health footer.
 
-### 13.2 CLI Detail Output
+### 13.3 CLI Detail Output
 
 File: `src/cli-output.ts`, `src/commands/helpers.ts` (`humanFormatWorkItem`)
 
@@ -322,7 +334,7 @@ Status:   🟢 [OPEN]
 Priority: 🔴 [CRIT]
 ```
 
-### 13.3 Tests
+### 13.4 Tests
 
 Tests should verify:
 - Icon functions return expected emoji for valid inputs
@@ -336,7 +348,7 @@ Tests should verify:
 ## 14. Appendix: Example Usage
 
 ```ts
-import { priorityIcon, statusIcon, iconsEnabled } from '../icons.js';
+import { priorityIcon, statusIcon, iconsEnabled } from './theme.js';
 
 const useIcons = iconsEnabled({ noIcons: opts.noIcons });
 
@@ -359,7 +371,8 @@ lines.push(`Priority: ${pIcon} ${item.priority}`);
 
 | File | Change |
 |------|--------|
-| `src/icons.ts` | Core icon module with emoji, fallback, and label functions |
+| `src/theme.ts` | Single source of truth: chalk colours + all icon maps/functions + `IconOptions` interface |
+| `src/icons.ts` | Deprecated re-export wrapper (re-exports from `theme.js`) — will be removed in WL-0MSJ4BT4Z002HH9B |
 | `src/commands/helpers.ts` | Added icon formatting to CLI output (summary, concise, normal, full) |
 | `src/commands/list.ts` | Added `--no-icons` CLI flag |
 | `src/commands/show.ts` | Added `--no-icons` CLI flag |
